+++ nessai/flows/utils.py
@@ -5,15 +5,16 @@ Various utilities for implementing normalising flows.
 import inspect
 import logging
 from typing import Optional, Type, Union
+import warnings
 
-from nflows import transforms
-from nflows.distributions import Distribution
+from glasflow.nflows import transforms
+from glasflow.nflows.distributions import Distribution
 import numpy as np
 import torch
 import torch.nn.functional as F
 
 from .distributions import MultivariateNormal
-from .transforms import LULinear
+
 
 logger = logging.getLogger(__name__)
 
@@ -40,7 +41,7 @@ def get_base_distribution(
     ----------
     n_inputs : int
         Number of inputs to the distribution.
-    distribution : Union[str, Type[nflows.distribution.Distribution]]
+    distribution : Union[str, Type[glasflow.nflows.distribution.Distribution]]
         Distribution class or name of known distribution
     kwargs : Any
         Keyword arguments used when creating an instance of distribution.
@@ -234,9 +235,9 @@ def reset_weights(module):
 
     Uses the ``reset_parameters`` method from ``torch.nn.Module``
 
-    Also checks the following modules from nflows
+    Also checks the following modules from glasflow.nflows
 
-    - nflows.transforms.normalization.BatchNorm
+    - glasflow.nflows.transforms.normalization.BatchNorm
 
     Parameters
     ----------
@@ -246,7 +247,7 @@ def reset_weights(module):
     if hasattr(module, "reset_parameters"):
         module.reset_parameters()
     elif isinstance(module, transforms.BatchNorm):
-        # nflows BatchNorm does not have a weight reset, so must
+        # glasflow.nflows BatchNorm does not have a weight reset, so must
         # be done manually
         constant = np.log(np.exp(1 - module.eps) - 1)
         module.unconstrained_weight.data.fill_(constant)
@@ -268,6 +269,9 @@ def reset_permutations(module):
     module : :obj:`torch.nn.Module`
         Module to reset
     """
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore")
+        from .transforms import LULinear
     if isinstance(module, (transforms.LULinear, LULinear)):
         module.cache.invalidate()
         module._initialize(identity_init=True)
@@ -291,7 +295,9 @@ def create_linear_transform(linear_transform, features):
         return transforms.CompositeTransform(
             [
                 transforms.RandomPermutation(features=features),
-                LULinear(features, identity_init=True, using_cache=True),
+                transforms.LULinear(
+                    features, identity_init=True, using_cache=True
+                ),
             ]
         )
     elif linear_transform.lower() == "svd":