+++ tests/test_deprecation_warnings.py
@@ -18,3 +18,12 @@ def test_nested_sampler_deprecation():
     with pytest.warns(FutureWarning) as record:
         from nessai import nestedsampler  # noqa
     assert "`nessai.nestedsampler` is deprecated" in str(record[0].message)
+
+
+def test_lulinear_warning():
+    """Assert a warning is raised when LULinear is imported"""
+    with pytest.warns(FutureWarning) as record:
+        from nessai.flows.transforms import LULinear  # noqa
+    assert "`nessai.flows.transforms.LULinear` is deprecated" in str(
+        record[0].message
+    )